third_party/blink/web_tests/fast/hidpi/image-srcset-fraction.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    addEventListener("load", function() {
        if (internals)
            shouldBeTrue('document.getElementById("testimg").clientWidth==200');
    }, false);
</script>
<div>This test passes if the srcset resource is loaded and displayed as the image</div>
<img src="resources/blue-100-px-square.png" srcset="resources/green-400-px-square.png 2x" id="testimg">
